cowebscraping/requirements.txt
Scrapy
python-dotenv
pymongo
unidecode
pytz
pyyaml

